requirements.txt
Faker
psycopg2-binary
confluent-kafka
pyflink
python-dotenv
fastapi
uvicorn
